+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/errorprone/InvalidLogMessageFormatRule.java
@@ -9,6 +9,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -36,6 +37,7 @@ import net.sourceforge.pmd.lang.java.symboltable.VariableNameDeclaration;
 import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import net.sourceforge.pmd.lang.symboltable.Scope;
+import net.sourceforge.pmd.util.StringUtil;
 
 public class InvalidLogMessageFormatRule extends AbstractJavaRule {
 
@@ -47,6 +49,12 @@ public class InvalidLogMessageFormatRule extends AbstractJavaRule {
             Pattern.compile("(\\{\\})|(%(?:\\d\\$)?(?:\\w+)?(?:\\d+)?(?:\\.\\d+)?\\w)");
 
     private static final Map<String, Set<String>> LOGGERS;
+
+    /**
+     * Whitelisted methods of net.logstash.logback.argument.StructuredArguments
+     */
+    private static final String[] STRUCTURED_ARGUMENTS_METHODS = {"a", "array", "defer", "e",
+        "entries", "f", "fields", "keyValue", "kv", "r", "raw", "v", "value"};
 
     static {
         Map<String, Set<String>> loggersMap = new HashMap<>();
@@ -116,14 +124,23 @@ public class InvalidLogMessageFormatRule extends AbstractJavaRule {
         final List<ASTExpression> argumentList = parentNode.getFirstChildOfType(ASTPrimarySuffix.class)
                 .getFirstDescendantOfType(ASTArgumentList.class).findChildrenOfType(ASTExpression.class);
 
-        if (argumentList.get(0).getType() != null && !argumentList.get(0).getType().equals(String.class)) {
-            if (argumentList.size() == 1) {
-                // no need to check for message params in case no string and no params found
-                return data;
+        // remove any arguments before the string message - these might be method calls for
+        // logstash markers or structured arguments
+        // this also removes any non-string value, e.g. a slf4j-Marker
+        // if the type cannot be determined, it is considered not to be a string...
+        Iterator<ASTExpression> iterator = argumentList.iterator();
+        while (iterator.hasNext()) {
+            ASTExpression argument = iterator.next();
+            if (!TypeTestUtil.isA(String.class, argument)) {
+                iterator.remove();
             } else {
-                // ignore the first argument if it is a known non-string value, e.g. a slf4j-Marker
-                argumentList.remove(0);
+                break;
             }
+        }
+
+        if (argumentList.isEmpty()) {
+            // no need to check for message params in case no string message found
+            return data;
         }
 
         // remove the message parameter
@@ -144,6 +161,12 @@ public class InvalidLogMessageFormatRule extends AbstractJavaRule {
         // But only, if it is not used as a placeholder argument
         if (argumentList.size() > expectedArguments) {
             removeThrowableParam(argumentList);
+        }
+        
+        // remove any logstash structured arguments at the end
+        // but only, if there are not enough placeholders
+        if (argumentList.size() > expectedArguments) {
+            removePotentialStructuredArguments(argumentList.size() - expectedArguments, argumentList);
         }
 
         int providedArguments = argumentList.size();
@@ -318,5 +341,51 @@ public class InvalidLogMessageFormatRule extends AbstractJavaRule {
             }
         }
         return stringLiterals;
+    }
+
+    /**
+     * Removes up to {@code maxArgumentsToRemove} arguments from the end of the {@code argumentList},
+     * if the argument is a method call to one of the whitelisted StructuredArguments methods.
+     *
+     * @param maxArgumentsToRemove
+     * @param argumentList
+     */
+    private void removePotentialStructuredArguments(int maxArgumentsToRemove, List<ASTExpression> argumentList) {
+        int removed = 0;
+        while (!argumentList.isEmpty() && removed < maxArgumentsToRemove) {
+            int lastIndex = argumentList.size() - 1;
+            ASTExpression argument = argumentList.get(lastIndex);
+            if (isStructuredArgumentMethodCall(argument)) {
+                argumentList.remove(lastIndex);
+                removed++;
+            } else {
+                // stop if something else is encountered
+                break;
+            }
+        }
+    }
+
+    /*
+     * └─ Expression == argument
+     *    └─ PrimaryExpression
+     *       ├─ PrimaryPrefix
+     *       │  └─ Name: eg. "keyValue" or "StructuredArguments.keyValue"
+     *       └─ PrimarySuffix
+     *          └─ Arguments
+     */
+    private boolean isStructuredArgumentMethodCall(ASTExpression argument) {
+        if (argument.getNumChildren() == 1 && argument.getChild(0) instanceof ASTPrimaryExpression) {
+            ASTPrimaryExpression primary = (ASTPrimaryExpression) argument.getChild(0);
+            if (primary.getNumChildren() == 2 && primary.getChild(1) instanceof ASTPrimarySuffix) {
+                ASTPrimaryPrefix prefix = (ASTPrimaryPrefix) primary.getChild(0);
+                ASTPrimarySuffix suffix = (ASTPrimarySuffix) primary.getChild(1);
+                if (suffix.isArguments() && prefix.getNumChildren() == 1 && prefix.getChild(0) instanceof ASTName) {
+                    ASTName name = (ASTName) prefix.getChild(0);
+                    return name.getImage().startsWith("StructuredArguments.")
+                        || StringUtil.isAnyOf(name.getImage(), STRUCTURED_ARGUMENTS_METHODS);
+                }
+            }
+        }
+        return false;
     }
 }